refactor: streamline entity imports and enhance MikroORM configuration
- Updated MikroORM configuration to use glob patterns for entity imports, improving compatibility with Vercel.
- Simplified entity imports across application modules by consolidating them into a single import from the entities directory.
- Enhanced clarity and maintainability of the codebase by reducing the number of direct entity imports.
