+++ src/entities/index.ts
@@ -1,0 +1,14 @@
+// Centralized entity exports for clean architecture
+export { User } from './models/user.entity';
+export { Todo } from './models/todo.entity';
+export { Session } from './models/session.entity';
+
+// Export model types
+export type { Cookie } from './models/cookie';
+
+// Export types as well for convenience
+export type * from './types';
+
+// Export error types
+export * from './errors/auth';
+export * from './errors/common';
